+++ app/[username]/npcs/[npcId]/page.tsx
@@ -1,7 +1,7 @@
 import {npcs} from '@/database/drizzle/schema';
 import {redirect} from 'next/navigation';
 import NPCCard from '@/components/cards/NPCCard';
-import {getUsername} from '@/server-actions/auth';
+import {getUsername} from '@/actions/auth';
 
 export default async function NPCDetailPage({
 	params,
