+++ app/page.tsx
@@ -1,6 +1,6 @@
 import Hero from '@/components/layout/Hero';
 import {redirect} from 'next/navigation';
-import {getUsername} from '@/server-actions/auth';
+import {getUsername} from '@/actions/auth';
 import SignUpForm from '@/components/forms/SignUpForm';
 
 export default async function Index() {
